backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pillow==10.4.0
# Optional for real NST; if unavailable, backend will fallback
torch>=2.2.0; platform_system!="Darwin" or platform_machine!="arm64"
torchvision>=0.17.0; platform_system!="Darwin" or platform_machine!="arm64"

numpy>=1.26.0

diffusers>=0.35.0
transformers>=4.41.2
accelerate>=0.27.0
flask>=3.0.0
flask_cors>=4.0.0